third_party/blink/web_tests/virtual/controls-refresh/color/color-picker-keyboard-up-down-space-format-toggle-expected.txt

Popup opened.
HSL
HEX
RGB
HEX
HSL
RGB
HSL
HEX
RGB
TEST COMPLETE
